requirements.txt
absl-py==0.13.0
chex==0.1.1
distrax==0.1.1
dm-haiku==0.0.6
dm-tree==0.1.6
jax==0.3.2
jaxlib==0.3.2
ml_collections==0.1.1
numpy==1.21.5
optax==0.1.0
tensorflow-probability==0.15.0
